[IMP] chart: smart chart 2.0 at chart insertion
Some improvements at the insertion of a new chart :

- if the selection is a single isolated cell, insert a scorecard
- if the selection is a single cell, try to expand the selection to find
a 'table' (contiguous non-empty cells)
- if the labels are all number/dates, insert a line chart
- else keep current behaviour, insert a bar chart

Odoo task 2916274
