+++ src/components/side_panel/chart/line_chart/line_chart_config_panel.ts
@@ -7,7 +7,7 @@ export class LineConfigPanel extends LineBarPieConfigPanel {
   get canTreatLabelsAsText() {
     const chart = this.env.model.getters.getChart(this.props.figureId);
     if (chart && chart instanceof LineChart) {
-      return canChartParseLabels(chart, this.env.model.getters);
+      return canChartParseLabels(chart.labelRange, this.env.model.getters);
     }
     return false;
   }
