+++ tests/menu_items_registry.test.ts
@@ -7,22 +7,12 @@ import {
   colMenuRegistry,
   createFullMenuItem,
   FullMenuItem,
-  MenuItemRegistry,
   rowMenuRegistry,
   topbarMenuRegistry,
 } from "../src/registries/index";
 import { getMenuChildren } from "../src/registries/menus/helpers";
 import { SpreadsheetChildEnv } from "../src/types";
 import {
-  BACKGROUND_CHART_COLOR,
-  DEFAULT_CELL_HEIGHT,
-  DEFAULT_CELL_WIDTH,
-  DEFAULT_FIGURE_HEIGHT,
-  DEFAULT_FIGURE_WIDTH,
-} from "./../src/constants";
-import {
-  addColumns,
-  addRows,
   copy,
   createFilter,
   freezeColumns,
@@ -37,8 +27,10 @@ import {
 } from "./test_helpers/commands_helpers";
 import { getCellContent } from "./test_helpers/getters_helpers";
 import {
+  doAction,
+  getName,
+  getNode,
   makeTestFixture,
-  mockChart,
   MockClipboard,
   mockUuidV4To,
   mountSpreadsheet,
@@ -47,39 +39,6 @@ import {
   target,
 } from "./test_helpers/helpers";
 jest.mock("../src/helpers/uuid", () => require("./__mocks__/uuid"));
-
-function getNode(
-  _path: string[],
-  menuRegistry: MenuItemRegistry = topbarMenuRegistry
-): FullMenuItem {
-  const path = [..._path];
-  const root = path.splice(0, 1)[0];
-  let node = menuRegistry.get(root);
-  for (let p of path) {
-    node = node.children
-      .filter((item): item is FullMenuItem => typeof item !== "function")
-      .find((child) => child.id === p)!;
-  }
-  return node;
-}
-
-function doAction(
-  path: string[],
-  env: SpreadsheetChildEnv,
-  menuRegistry: MenuItemRegistry = topbarMenuRegistry
-): void {
-  const node = getNode(path, menuRegistry);
-  node.action(env);
-}
-
-function getName(
-  path: string[],
-  env: SpreadsheetChildEnv,
-  menuRegistry: MenuItemRegistry = topbarMenuRegistry
-): string {
-  const node = getNode(path, menuRegistry);
-  return typeof node.name === "function" ? node.name(env).toString() : node.name.toString();
-}
 
 describe("Menu Item Registry", () => {
   let menuDefinitions;
@@ -1048,385 +1007,6 @@ describe("Menu Item actions", () => {
     expect(dispatch).toHaveBeenCalledWith("SET_GRID_LINES_VISIBILITY", {
       sheetId,
       areGridLinesVisible: false,
-    });
-  });
-
-  describe("Insert > Chart", () => {
-    const data = {
-      sheets: [
-        {
-          name: "Sheet1",
-          rows: {},
-          cells: {
-            A2: { content: "P1" },
-            A3: { content: "P2" },
-            A4: { content: "P3" },
-            A5: { content: "P4" },
-
-            B1: { content: "first column dataset" },
-            B2: { content: "10" },
-            B3: { content: "11" },
-            B4: { content: "12" },
-            B5: { content: "13" },
-
-            C1: { content: "" },
-            C2: { content: "2" },
-            C3: { content: "4" },
-            C4: { content: "6" },
-
-            D1: { content: "=sum()" },
-            D2: { content: "3" },
-            D3: { content: "2" },
-            D4: { content: "5" },
-
-            E1: { content: "Title1" },
-            E2: { content: "10" },
-            E3: { content: "11" },
-            E4: { content: "12" },
-
-            F1: { content: "=sum(1,2)" },
-            F2: { content: "7" },
-            F3: { content: "8" },
-            F4: { content: "9" },
-
-            G1: { content: "" },
-            G2: { content: "7" },
-            G3: { content: "8" },
-            G4: { content: "9" },
-
-            H1: { content: "Title2" },
-            H2: { content: "7" },
-            H3: { content: "8" },
-            H4: { content: "9" },
-          },
-        },
-      ],
-    };
-    let dispatchSpy: jest.SpyInstance;
-    let defaultPayload: any;
-
-    beforeEach(async () => {
-      fixture = makeTestFixture();
-      ({ app, parent, model } = await mountSpreadsheet(fixture, { model: new Model(data) }));
-      env = parent.env;
-      mockChart();
-      dispatchSpy = spyDispatch(parent);
-      defaultPayload = {
-        position: expect.any(Object),
-        size: expect.any(Object),
-        id: expect.any(String),
-        sheetId: model.getters.getActiveSheetId(),
-        definition: {
-          dataSets: ["A1"],
-          background: BACKGROUND_CHART_COLOR,
-          dataSetsHaveTitle: false,
-          labelRange: undefined,
-          legendPosition: "none",
-          stacked: false,
-          title: expect.any(String),
-          type: "bar",
-          verticalAxisPosition: "left",
-        },
-      };
-    });
-
-    afterEach(() => {
-      app.destroy();
-    });
-
-    test("Chart is inserted at correct position", () => {
-      setSelection(model, ["B2"]);
-      doAction(["insert", "insert_chart"], env);
-      const { width, height } = model.getters.getSheetViewDimension();
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2"];
-      payload.position = {
-        x: (width - DEFAULT_FIGURE_WIDTH) / 2,
-        y: (height - DEFAULT_FIGURE_HEIGHT) / 2,
-      }; // Position at the center of the viewport
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart is selected and focused at insertion", async () => {
-      setSelection(model, ["B2"]);
-      doAction(["insert", "insert_chart"], env);
-      const id = model.getters.getFigures(model.getters.getActiveSheetId())[0].id;
-      expect(dispatchSpy).toHaveBeenCalledWith("SELECT_FIGURE", { id });
-      await nextTick();
-      expect(document.activeElement?.classList).toContain("o-figure");
-    });
-
-    test("Chart is inserted at correct position inside bottomRight pane for rows freeze", () => {
-      const sheetId = model.getters.getActiveSheetId();
-      freezeRows(model, 5, sheetId);
-      setSelection(model, ["B2"]);
-      doAction(["insert", "insert_chart"], env);
-      const { width, height } = model.getters.getSheetViewDimension();
-      const { y: offsetCorrectionY } = model.getters.getMainViewportCoordinates();
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2"];
-      payload.position = {
-        x: (width - DEFAULT_FIGURE_WIDTH) / 2,
-        y: (height - DEFAULT_FIGURE_HEIGHT + offsetCorrectionY) / 2,
-      }; // Position at the center of the viewport
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart is inserted at correct position inside bottomRight pane for columns freeze", () => {
-      const sheetId = model.getters.getActiveSheetId();
-      freezeColumns(model, 4, sheetId);
-      setSelection(model, ["B2"]);
-      doAction(["insert", "insert_chart"], env);
-      const { width, height } = model.getters.getSheetViewDimension();
-      const { x: offsetCorrectionX } = model.getters.getMainViewportCoordinates();
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2"];
-      payload.position = {
-        x: (width - DEFAULT_FIGURE_WIDTH + offsetCorrectionX) / 2,
-        y: (height - DEFAULT_FIGURE_HEIGHT) / 2,
-      }; // Position at the center of the viewport
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart is inserted at correct position inside bottomRight pane for both freeze", () => {
-      const sheetId = model.getters.getActiveSheetId();
-      freezeColumns(model, 4, sheetId);
-      freezeRows(model, 5, sheetId);
-      setSelection(model, ["B2"]);
-      doAction(["insert", "insert_chart"], env);
-      const { width, height } = model.getters.getSheetViewDimension();
-      const { x: offsetCorrectionX, y: offsetCorrectionY } =
-        model.getters.getMainViewportCoordinates();
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2"];
-      payload.position = {
-        x: (width - DEFAULT_FIGURE_WIDTH + offsetCorrectionX) / 2,
-        y: (height - DEFAULT_FIGURE_HEIGHT + offsetCorrectionY) / 2,
-      }; // Position at the center of the viewport
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart is inserted at the top left of the viewport when too small", () => {
-      setSelection(model, ["B2"]);
-      model.dispatch("RESIZE_SHEETVIEW", {
-        width: DEFAULT_FIGURE_WIDTH / 2,
-        height: DEFAULT_FIGURE_HEIGHT / 2,
-        gridOffsetX: 0,
-        gridOffsetY: 0,
-      });
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2"];
-      payload.position = {
-        x: 0,
-        y: 0,
-      }; // Position at the center of the viewport
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart is inserted at the top left of the viewport when too small in a frozen pane", () => {
-      addRows(model, "before", 0, 100);
-      setSelection(model, ["B2"]);
-      model.dispatch("RESIZE_SHEETVIEW", {
-        width: DEFAULT_FIGURE_WIDTH * 1.5,
-        height: DEFAULT_FIGURE_HEIGHT * 1.5,
-        gridOffsetX: 0,
-        gridOffsetY: 0,
-      });
-      const { bottom, right } = model.getters.getActiveMainViewport();
-      freezeColumns(model, Math.floor(right / 2));
-      freezeRows(model, Math.floor(bottom / 2));
-      const { x: offsetCorrectionX, y: offsetCorrectionY } =
-        model.getters.getMainViewportCoordinates();
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2"];
-
-      payload.position = {
-        x: offsetCorrectionX,
-        y: offsetCorrectionY,
-      }; // Position at the top of the bottom pane of the viewport
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart is inserted at correct position on a scrolled viewport", () => {
-      setSelection(model, ["B2"]);
-      const { width, height } = env.model.getters.getSheetViewDimension();
-      addColumns(model, "after", "D", 100);
-      addRows(model, "after", 4, 100);
-      env.model.dispatch("SET_VIEWPORT_OFFSET", {
-        offsetX: 2 * DEFAULT_CELL_WIDTH,
-        offsetY: 4 * DEFAULT_CELL_HEIGHT,
-      });
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2"];
-      payload.position = {
-        x: 2 * DEFAULT_CELL_WIDTH + (width - DEFAULT_FIGURE_WIDTH) / 2,
-        y: 4 * DEFAULT_CELL_HEIGHT + (height - DEFAULT_FIGURE_HEIGHT) / 2,
-      }; // Position at the center of the viewport
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart is inserted at correct position on a scrolled viewport with frozen rows", () => {
-      const sheetId = model.getters.getActiveSheetId();
-      freezeRows(model, 5, sheetId);
-      setSelection(model, ["B2"]);
-      const { width, height } = model.getters.getSheetViewDimension();
-      const { y: offsetCorrectionY } = model.getters.getMainViewportCoordinates();
-      addColumns(model, "after", "D", 100);
-      addRows(model, "after", 4, 100);
-      env.model.dispatch("SET_VIEWPORT_OFFSET", {
-        offsetX: 2 * DEFAULT_CELL_WIDTH,
-        offsetY: 4 * DEFAULT_CELL_HEIGHT,
-      });
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2"];
-      payload.position = {
-        x: 2 * DEFAULT_CELL_WIDTH + (width - DEFAULT_FIGURE_WIDTH) / 2,
-        y: 4 * DEFAULT_CELL_HEIGHT + (height - DEFAULT_FIGURE_HEIGHT + offsetCorrectionY) / 2,
-      }; // Position at the center of the viewport
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart is inserted at correct position on a scrolled viewport with columns frozen", () => {
-      const sheetId = model.getters.getActiveSheetId();
-      freezeColumns(model, 4, sheetId);
-      setSelection(model, ["B2"]);
-      const { width, height } = model.getters.getSheetViewDimension();
-      const { x: offsetCorrectionX } = model.getters.getMainViewportCoordinates();
-      addColumns(model, "after", "D", 100);
-      addRows(model, "after", 4, 100);
-      env.model.dispatch("SET_VIEWPORT_OFFSET", {
-        offsetX: 2 * DEFAULT_CELL_WIDTH,
-        offsetY: 4 * DEFAULT_CELL_HEIGHT,
-      });
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2"];
-      payload.position = {
-        x: 2 * DEFAULT_CELL_WIDTH + (width - DEFAULT_FIGURE_WIDTH + offsetCorrectionX) / 2,
-        y: 4 * DEFAULT_CELL_HEIGHT + (height - DEFAULT_FIGURE_HEIGHT) / 2,
-      }; // Position at the center of the viewport
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart is inserted at correct position on a scrolled viewport with both directions frozen", () => {
-      const sheetId = model.getters.getActiveSheetId();
-      freezeColumns(model, 4, sheetId);
-      freezeRows(model, 5, sheetId);
-      setSelection(model, ["B2"]);
-      const { width, height } = model.getters.getSheetViewDimension();
-      const { x: offsetCorrectionX, y: offsetCorrectionY } =
-        model.getters.getMainViewportCoordinates();
-      addColumns(model, "after", "D", 100);
-      addRows(model, "after", 4, 100);
-      env.model.dispatch("SET_VIEWPORT_OFFSET", {
-        offsetX: 2 * DEFAULT_CELL_WIDTH,
-        offsetY: 4 * DEFAULT_CELL_HEIGHT,
-      });
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2"];
-      payload.position = {
-        x: 2 * DEFAULT_CELL_WIDTH + (width - DEFAULT_FIGURE_WIDTH + offsetCorrectionX) / 2,
-        y: 4 * DEFAULT_CELL_HEIGHT + (height - DEFAULT_FIGURE_HEIGHT + offsetCorrectionY) / 2,
-      }; // Position at the center of the viewport
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart of single column without title", () => {
-      setSelection(model, ["B2:B5"]);
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2:B5"];
-      payload.definition.labelRange = undefined;
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart of single column with title", () => {
-      setSelection(model, ["B1:B5"]);
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B1:B5"];
-      payload.definition.labelRange = undefined;
-      payload.definition.dataSetsHaveTitle = true;
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart of several columns (ie labels) without title", () => {
-      setSelection(model, ["A2:B5"]);
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B2:B5"];
-      payload.definition.labelRange = "A2:A5";
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart of several columns (ie labels) with title", () => {
-      setSelection(model, ["A1:B5"]);
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B1:B5"];
-      payload.definition.labelRange = "A2:A5";
-      payload.definition.dataSetsHaveTitle = true;
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-
-    test("Chart title should be set by default if dataset have any", () => {
-      setSelection(model, ["B1:C4"]);
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["C1:C4"];
-      payload.definition.legendPosition = "none";
-      payload.definition.title = "";
-      payload.definition.dataSetsHaveTitle = false;
-      payload.definition.labelRange = "B1:B4";
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-    test("Chart title should only generate string and numerical values", async () => {
-      setSelection(model, ["C1:G4"]);
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["D1:G4"];
-      payload.definition.legendPosition = "top";
-      payload.definition.title = "Title1 and 3";
-      payload.definition.dataSetsHaveTitle = true;
-      payload.definition.labelRange = "C2:C4";
-      await nextTick();
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-    test("Chart title should only append and prefix to last title", () => {
-      setSelection(model, ["C1:H4"]);
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["D1:H4"];
-      payload.definition.legendPosition = "top";
-      payload.definition.title = "Title1, 3 and Title2";
-      payload.definition.dataSetsHaveTitle = true;
-      payload.definition.labelRange = "C2:C4";
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-    test("[Case 1] Chart is inserted with proper legend position", () => {
-      setSelection(model, ["A1:B5"]);
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["B1:B5"];
-      payload.definition.labelRange = "A2:A5";
-      payload.definition.dataSetsHaveTitle = true;
-      payload.definition.legendPosition = "none";
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
-    });
-    test("[Case 2] Chart is inserted with proper legend position", () => {
-      setSelection(model, ["F1:I5"]);
-      doAction(["insert", "insert_chart"], env);
-      const payload = { ...defaultPayload };
-      payload.definition.dataSets = ["G1:I5"];
-      payload.definition.dataSetsHaveTitle = true;
-      payload.definition.labelRange = "F2:F5";
-      payload.definition.legendPosition = "top";
-      expect(dispatchSpy).toHaveBeenCalledWith("CREATE_CHART", payload);
     });
   });
